Add 'init' task in Makefile and example local environment file
A new task called 'init' was added to the Makefile that aims to setup the local environment for new developers. This task will copy the .env.local.example file to .env.local if it does not exist already. This is designed to streamline and simplify the project setup process, making it more approachable for newcomers.