Add Conference entity and related migration
New Conference entity has been added to the system. This entity has fields like city, year, and isInternational. To make this change possible at the database level, a new Doctrine migration is also added. This migration handles the creation of related conference table and fields in up method and dropping them in down method. In addition, ConferenceRepository is created for future data handling of Conference entity. This has been done to support management of conferences in the application.