Add conference relation to comment entity
This commit adds a ManyToOne relationship from the Comment entity to the Conference entity. As a result, the feature that retrieves and assigns the conference of each comment is provided.  Also, this commit introduces a database migration required for the change. This was necessary to link the comments to their respective conference. This improves data structure and enables more granular data manipulation.