Add photo filename field to Comment entity
This commit introduces a new field, 'photo_filename', to the Comment entity and its respective Doctrine migration files.
It's necessary to store the reference image filename associated with each comment for further usage. The new attribute is being added in the Comment.php and Version20230714195125.php migration file for maintaining Database versioning.