Add EasyAdmin bundle and Symfony UID components
Added the 'easycorp/easyadmin-bundle' and 'symfony/uid' to the 'composer.json' file to facilitate the creation of an admin interface and to provide an object-oriented API to generate and represent UIDs respectively. Bundle registration in 'config/bundles.php' and UID configuration in 'config/packages/uid.yaml' were also updated accordingly. Changes in the 'composer.lock', 'symfony.lock', and additions to other components are a consequence of the EasyAdmin bundle requirements.