Make photoFilename nullable in Comment Entity
A change was made to the Comment Entity class to set the column 'photoFilename' as nullable. Previously, this field was required, but after analysing system usage, it was understood that not every comment necessarily has to have a photo attached. A new migration file was also created to reflect this change in the database schema.