Update session handler and database migrations
In this commit, the session handler in framework.yaml is updated to read from DATABASE_URL instead of being null, enabling a centralized session store in a database. Moreover, a new migration file (Version20230715103943.php) is created for creating 'sessions' table to store the session data.

The unnecessary command of 'CREATE SCHEMA public' has been removed from the down method in the migration file 'Version20230714205334.php'. 'CREATE SCHEMA public' was found to be irrelevant since we are not creating any new schema.