Refactor Conference fetching logic for better performance
Changes were made to the ConferenceRepository and ConferenceController classes to enhance the efficiency of data fetching. The findAll() method in the ConferenceRepository is now sorting Conferences by 'year' and 'city' right at the fetch time. Consequently, ConferenceRepository was removed from all parts of ConferenceController and a new file, TwigEventSubscriber, was introduced. The conferences are now being fetched globally just once per request, which should improve the performance.