Update Comment Entity, auto-assign 'createdAt' value on creation
In the Comment entity (src/Entity/Comment.php), an ORM lifecycle callback with a prePersist annotation was added to automatically assign the current date and time to the 'createdAt' property right before the entity gets persisted. The DateTimeImmutable object is used for this purpose.
Also, in the Comment CRUD Controller (src/Controller/Admin/CommentCrudController.php) 'createdAt' field is now disabled when editing a comment. This ensures the original creation time remains intact and reduces the error scope by preventing manual changes.