Add slug computation to Conference Entity
A new method named "computeSlug" has been added to the Conference entity class in order to automatically generate Conference slugs. This ensures that the slug property of a Conference object will have a proper value before persisting or updating it in the database. SluggerInterface from Symfony Components was used for this purpose. Furthermore, the ConferenceEntityListener class has been created to call the computeSlug method before persisting or updating any Conference object. Additionally, constructors in TwigEventSubscriber class were modified to be more simplifying and readable.