Add photo uploading feature to comments
Added a feature allowing users to upload photos along with their comments. This feature can add more depth and context to comment threads. It also includes form validation to ensure that only valid images are uploaded. This required changes to the .gitignore file, adding a new path to store the user-uploaded images. The services.yaml file was modified to include the path of the photos as a parameter. Corresponding changes were also made to the ConferenceController to handle photo uploads and updates to the comment entity and form to include photo. In addition, the show.html.twig template was updated to include the photo upload form field.