Fix conference path hyperlink in templates
The commit removes the hard-coded "/" in the link for each conference displayed on the conference index page. Instead of hard-coding the "/," the Twig path function is used to dynamically generate the URL. This change leads to more flexible code, allowing the URLs of the site to change without having to manually update this link.