Add migration command and update README
Extended the PHONY list in Makefile to include 'migration' and 'migrate' commands for a smoother database management experience. The 'make-migration' command has been renamed to 'migration' for conciseness. README is updated to better guide users through the process of setup. The URL and default login credentials for the Admin page are also included.