Add Akismet key to environment variables and secrets configuration
Akismet key is added to the .env.local.example file and references to it were added in the config/secrets files for both dev and prod environments. The key was also added to the services.yaml file in the parameters section. This addition is necessary for enabling Akismet services, which are used to filter out spam content in the application.