Add testing commands to Makefile and testing secrets
Added `init-tests`, `composer`, and `tests` commands to the Makefile for initializing and running tests on the application inside the Docker container. Updated `migration` and `migrate` commands to include a no-interaction option.

Also, added encryption-specific testing secrets to the Symfony configuration in order to securely handle test secrets. The new commands and secrets are necessary for creating and configuring a testing environment for the application.