Add "doctrine/doctrine-fixtures-bundle" to support dev/testing
This commit adds the "doctrine/doctrine-fixtures-bundle" library to support development and testing. The bundle helps in managing the data fixtures for the Doctrine2 ORM. This is beneficial for the tests which require specific data setup. Dev/test environments are ensured with the tools necessary for effective testing. Changes are reflected in composer.json, composer.lock, symfony.lock, and config/bundles.php files. A skeleton file for the application's DataFixtures was also created.