Add publish state to Comment entity
A new property 'state' has been added to the Comment entity to track the publishing status of each comment. This property is of type PublishState Enum which includes 'submitted', 'spam', 'published'. The CommentRepository has been updated to only fetch comments that have 'published' state. The CommentCrudController is also updated to show the state in the admin UI. This change will help in better management and classification of comments.