Add publish state to comment fixture
In the AppFixtures file, I added a new property for the Comment fixture to simulate the state of a comment. This was achieved by including the PublishState Enum and setting it in the Comment fixture via the getWeightedPublishStates method which provides a weighted list of publish states (Submitted, Spam, Published). This change brings the fixtures closer to a real-life scenario where comments are in various states of publish.