Add AMQP Messenger for message queuing
This commit introduces the Symfony AMQP Messenger package which will enable our application to handle asynchronous messaging. This is beneficial as it can improve application performance by queuing tasks to be handled in the background. The Docker configuration has also been updated to add RabbitMQ as a message broker. The ".env" file was updated to change the MESSENGER_TRANSPORT_DSN to use AMQP.