Update PHP Dockerfile and Docker-compose settings
Updated PHP Dockerfile to more configurable setup and changed Docker Compose files to include an 'app-consumer' service. This allows different PHP types to be set, and enables the application to consume messages from a message broker, improving functionality. Additionally, Dockerfile now sets the working directory to /var/www/html to follow convention.

In the .env file, the MESSENGER_TRANSPORT_DSN variable was updated to 'rabbit' to correctly point to the RabbitMQ service in Docker and removed unnecessary options in messenger.yaml for a cleaner configuration. The message-broker service in docker-compose.yml was updated to include ports, image was changed to RabbitMQ management for easier debugging, and dependencies were set for app and app-consumer services.