Update healthcheck interval and messenger watch command
Updated the health check parameters for RabbitMQ service in docker-compose.override.yml to make it more responsive. The interval and timeout has been reduced to 1s and number of retries increased to 30. This will ensure quicker detection of any issues with the service.

Also adjusted the messenger:consume command in docker-compose.yml to watch changes in the 'config', 'src', 'templates' and 'vendor' directories. This modification will automatically restart the consumer when related files are updated, improving the development workflow.