Enable Edge Side Includes and add a cache for conference header
Enabled the Edge Side Includes (ESI) in the framework's config file to enable more efficient cache management. By shifting the iteration over the conferences into the conference controller we can now cache the conference header independently for better performance. This required the creation of a separate 'conference_header' function and corresponding route, and thus a new Twig template for the header. Also, a subscriber file was deleted as it became unnecessary with these changes.