Add Node.js and npm to Dockerfile and update Makefile
In the Dockerfile, Node.js and npm have been added to the list of packages installed by apk. This is to support asset compilation inside the dockerized environment.

In the Makefile, two commands 'watch-assets' and 'clear-cache' have been added. The 'watch-assets' command is for compiling and reloading assets during development. The 'clear-cache' command clears Symfony's cache, improving performance.