Add Image Optimizer service and integrate within workflow
ImageOptimizer service to resize user uploaded photos has been added to manage memory usage efficiently. It has been configured in the workflow to run before each image is published. This optimizes the performance and load times of the application, improving the user experience. It further assists in efficient disk space utilization. Extensions including 'imagine/imagine' were added to composer for this purpose. The namespace 'app' is introduced and used in yaml files to avoid naming collisions. A new status 'ready' has been introduced in the publishing workflow to indicate that an image is resized and ready for publishing.