Add CommentCleanupCommand and refactor codebase
New CommentCleanupCommand has been added to manage the removal process of rejected and spam comments from the database. Also, a small typo has been fixed in the description of the StepInfoCommand. Additionally, methods to count and delete old rejected comments have been added to CommentRepository. The purpose of these changes is to enhance code usability and maintainability.