Refactor comment notification system
The notification system in CommentMessageHandler has been refactored to use the NotifierInterface instead of the MailerInterface. This change simplifies the process of sending review notifications by introducing the CommentReviewNotification class. It also includes the extraction of email configuration to a YAML file for easier management. This allows administrators to set up the alert system in a more streamlined and efficient manner.