Add symfony/fake-chat-notifier package for development
The symfony/fake-chat-notifier package was added to the composer.json and was registered in the .env and notifier.yaml config files. This package will allow for testing notifications during development without sending actual requests to the chat services. Updates were also made to the CommentReviewNotification class to make use of this new notifier. The commit also includes changes to the composer.lock and symfony.lock files reflecting the new package addition.