Add reviewUrl to comment notification
This commit introduces a link in the review notifications, pointing directly to the comments to be reviewed. The UrlGeneratorInterface was used to create a fully qualified URL for the comment review. The inclusion of this link will streamline the review process by enabling reviewers to access the comments directly from the notification, thus improving the user experience.