Add NelmioCorsBundle and ApiPlatformBundle
Added NelmioCorsBundle for managing Cross-Origin Resource Sharing (CORS) and the ApiPlatformBundle for creating APIs. This includes installing new packages, configuring CORS settings, and adding routes for API Platform. The NelmioCorsBundle allows the restriction of CORS headers to specified origins, which is crucial for restricting which websites can access the application APIs. Similarly, the addition of API Platform helps in quickly building out REST and GraphQL APIs.