Add RABBITMQ_URL to .env and update MESSENGER_TRANSPORT_DSN
This commit includes two changes in the .env file. First, a new variable RABBITMQ_URL was declared to store the URL for the RabbitMQ message broker. This changes provide a way to abstain from hardcoding the RabbitMQ URL directly into the MESSENGER_TRANSPORT_DSN. Secondly, the MESSENGER_TRANSPORT_DSN was updated to utilize the newly introduced RABBITMQ_URL, making the app more configurable and cleaner.